refactor: introduce service layer for business logic separation
