+++ src/routes/api/settings/-display.ts
@@ -1,27 +1,16 @@
 /**
  * Display settings API endpoints.
- * Provides GET and PUT operations for user display preferences (sidebar items).
+ * Delegates to the display service for business logic.
  */
 
 import { Elysia, t } from "elysia";
-import { eq } from "drizzle-orm";
 import { auth } from "~/lib/auth";
-import { db, schema } from "~/lib/db";
-import { createError } from "evlog";
-import { nanoid } from "nanoid";
-import { settingsLogger } from "~/lib/logger";
+import { getDisplay, updateDisplay } from "~/services/settings";
 
-/**
- * Display data schema for OpenAPI documentation.
- */
 const displayExample = {
   items: ["recents", "home"],
 };
 
-/**
- * Display settings route group.
- * Mounted under `/api/settings/display`.
- */
 export const displaySettingsRoutes = new Elysia({
   name: "api.routes.settings.display",
   prefix: "/settings/display",
@@ -30,34 +19,12 @@ export const displaySettingsRoutes = new Elysia({
     "/",
     async ({ set, request }) => {
       const session = await auth.api.getSession({ headers: request.headers });
-
       if (!session) {
         set.status = 401;
-        return createError({
-          message: "Unauthorized",
-          status: 401,
-          why: "No active session found",
-          fix: "Sign in to access your display settings",
-        });
+        return { error: "Unauthorized" };
       }
-
-      const userId = session.user.id;
-
-      const [display] = await db
-        .select()
-        .from(schema.userSettingsDisplay)
-        .where(eq(schema.userSettingsDisplay.userId, userId))
-        .limit(1);
-
-      if (!display) {
-        return {
-          items: ["recents", "home"],
-        };
-      }
-
-      return {
-        items: JSON.parse(display.items || '["recents","home"]'),
-      };
+      const data = await getDisplay(session.user.id);
+      return data;
     },
     {
       detail: {
@@ -78,53 +45,14 @@ export const displaySettingsRoutes = new Elysia({
     "/",
     async ({ body, set, request }) => {
       const session = await auth.api.getSession({ headers: request.headers });
-
       if (!session) {
         set.status = 401;
-        return createError({
-          message: "Unauthorized",
-          status: 401,
-          why: "No active session found",
-          fix: "Sign in to update your display settings",
-        });
+        return { error: "Unauthorized" };
       }
 
-      const userId = session.user.id;
       const { items } = body as { items: string[] };
-
-      const existing = await db
-        .select()
-        .from(schema.userSettingsDisplay)
-        .where(eq(schema.userSettingsDisplay.userId, userId))
-        .limit(1);
-
-      const now = new Date();
-
-      if (existing.length > 0) {
-        await db
-          .update(schema.userSettingsDisplay)
-          .set({
-            items: JSON.stringify(items || ["recents", "home"]),
-            updatedAt: now,
-          })
-          .where(eq(schema.userSettingsDisplay.userId, userId));
-
-        settingsLogger.debug("Display updated", { userId });
-      } else {
-        await db.insert(schema.userSettingsDisplay).values({
-          id: nanoid(),
-          userId,
-          items: JSON.stringify(items || ["recents", "home"]),
-          createdAt: now,
-          updatedAt: now,
-        });
-
-        settingsLogger.debug("Display created", { userId });
-      }
-
-      return {
-        items: items || ["recents", "home"],
-      };
+      const data = await updateDisplay(session.user.id, { items });
+      return data;
     },
     {
       body: t.Object({
