+++ src/services/settings/index.ts
@@ -1,0 +1,4 @@
+export * from "./profile";
+export * from "./account";
+export * from "./display";
+export * from "./notifications";
